Seperate BrewManager into utils
